refactor: Worksheet 用 TableCell の切り出し
