+++ lib/presentation/diagnosis/table_page.dart
@@ -6,6 +6,7 @@ import 'package:mood_trend_flutter/utils/app_colors.dart';
 
 import '../../application/diagnosis/states/subscribe_mood_work_sheet_provider.dart';
 import '../../utils/page_navigator.dart';
+import 'components/worksheet_table_cell.dart';
 import 'manic/manic_type_diagnosis_page.dart';
 
 /// [MoodState] は、気分値目安表の状態を表す
@@ -119,7 +120,7 @@ class TablePage extends ConsumerWidget {
                               ref.watch(selectedMoodButtonStateProvider) ==
                                       MoodState.manic
                                   ? [
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '+5',
                                         actionText: worksheet.plus_5,
                                       ),
@@ -130,7 +131,7 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '+4',
                                         actionText: worksheet.plus_4,
                                       ),
@@ -141,7 +142,7 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '+3',
                                         actionText: worksheet.plus_3,
                                       ),
@@ -152,7 +153,7 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '+2',
                                         actionText: worksheet.plus_2,
                                       ),
@@ -163,13 +164,13 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '+1',
                                         actionText: worksheet.plus_1,
                                       ),
                                     ]
                                   : [
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '-1',
                                         actionText: worksheet.minus_1,
                                       ),
@@ -180,7 +181,7 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '-2',
                                         actionText: worksheet.minus_2,
                                       ),
@@ -191,7 +192,7 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '-3',
                                         actionText: worksheet.minus_3,
                                       ),
@@ -202,7 +203,7 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '-4',
                                         actionText: worksheet.minus_4,
                                       ),
@@ -213,7 +214,7 @@ class TablePage extends ConsumerWidget {
                                         endIndent: 16,
                                         color: AppColors.white,
                                       ),
-                                      TableCell(
+                                      WorksheetTableCell(
                                         moodValue: '-5',
                                         actionText: worksheet.minus_5,
                                       ),
@@ -248,56 +249,6 @@ class TablePage extends ConsumerWidget {
           ),
         );
       },
-    );
-  }
-}
-
-/// 気分値目安表のセル
-class TableCell extends StatelessWidget {
-  const TableCell({
-    super.key,
-    required this.moodValue,
-    required this.actionText,
-  });
-  final String moodValue;
-  final String actionText;
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Text(
-            moodValue,
-            style: const TextStyle(fontSize: 24),
-          ),
-        ),
-        Expanded(
-          child: SizedBox(
-            // これがないと文字数が多い時にUIが崩れる
-            height: 80,
-            child: Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Center(
-                child: Row(
-                  children: [
-                    Expanded(
-                      child: Text(
-                        textAlign: TextAlign.left,
-                        actionText,
-                        style: TextStyle(
-                          fontSize: 16,
-                          color: AppColors.black,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
     );
   }
 }
